+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 4.1.0 - 2025-03-21
+
+### Added
+
+- Support for `innmind/black-box` `6`
+
+### Fixed
+
+- Support for PHP `8.4`
+
 ## 4.0.0 - 2024-07-14
 
 ### Added